requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorthief==0.2.1
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-PyMongo==3.0.1
fonttools==4.57.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.11.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
urllib3==2.3.0
webcolors==24.11.1
Werkzeug==3.1.3
